move functions to keeper.go
